clear the screenshots folder; move all screenshot-related code into screenshots.go
