+++ script/script.go
@@ -2,10 +2,7 @@ package script
 
 import (
 	"context"
-	"fmt"
 	"log"
-	"os/exec"
-	"path/filepath"
 	"time"
 
 	"github.com/ftl/hellocontest/core/app"
@@ -53,28 +50,5 @@ func Wait(duration time.Duration) Step {
 	return func(context.Context, *app.Controller, func(func())) time.Duration {
 		log.Printf("[WAITING FOR %v]", duration)
 		return duration
-	}
-}
-
-func TriggerScreenshot(filename string) Step {
-	return TriggerScreenshotWithDelay(filename, 0)
-}
-
-func TriggerScreenshotWithDelay(filename string, delay time.Duration) Step {
-	return func(_ context.Context, _ *app.Controller, _ func(func())) time.Duration {
-		// TODO: evaluate ctx.Done() and stop the flameshot process
-		cmd := exec.Command("flameshot", "gui")
-		cmd.Args = append(cmd.Args, "--path", filepath.Join("./docs/screenshots", filename+".png"))
-		if delay > 0 {
-			cmd.Args = append(cmd.Args, "--delay", fmt.Sprintf("%d", delay.Milliseconds()))
-		}
-
-		err := cmd.Run()
-		if err != nil {
-			log.Printf("Screenshot failed: %v", err)
-		} else {
-			log.Println("Screenshot successful")
-		}
-		return 0
 	}
 }
